refactor: move date range variables to new file
